refactor(sponsor-reports): extract ContentCell into a shared component
